backend/pyproject.toml
[project]
name = "pm-backend"
version = "0.1.0"
description = "FastAPI backend scaffold for the PM MVP"
requires-python = ">=3.12"
dependencies = [
  "bcrypt>=4.2.0",
  "fastapi>=0.116.0",
  "mysql-connector-python>=9.0.0",
  "openai>=1.75.0",
  "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.4.0",
]

[tool.uv]
package = false
